+++ packages/react-static/src/browser/components/Routes.js
@@ -12,6 +12,14 @@ import { getRoutePath, isSSR } from '../utils'
 import onLocation from '../utils/Location'
 import Spinner from './Spinner'
 
+const RoutePathContext = React.createContext()
+
+export const withRoutePathContext = Comp => props => (
+  <RoutePathContext.Consumer>
+    {routePath => <Comp {...props} routePath={routePath} />}
+  </RoutePathContext.Consumer>
+)
+
 export default withStaticInfo(
   class Routes extends Component {
     static defaultProps = {
@@ -34,7 +42,7 @@ export default withStaticInfo(
     render() {
       const { children, Loader, staticInfo } = this.props
 
-      const currentRoutePath = isSSR() ? staticInfo.path : getCurrentRoutePath()
+      const routePath = isSSR() ? staticInfo.path : getCurrentRoutePath()
 
       const getComponentForPath = routePath => {
         // Clean the path
@@ -55,6 +63,11 @@ export default withStaticInfo(
         // Detect an unloaded template
         // TODO:suspense - This will become a suspense resource
         if (!Comp) {
+          if (is404) {
+            throw new Error(
+              'This page template could not be found and a 404 template could not be found to fall back on. This means something is terribly wrong and you should probably file an issue!'
+            )
+          }
           ;(async () => {
             await Promise.all([
               prefetch(routePath, { priority: true }),
@@ -73,14 +86,20 @@ export default withStaticInfo(
           ) : null
       }
 
-      if (children) {
-        return children({
-          getComponentForPath,
-        })
-      }
+      const Comp = getComponentForPath(routePath)
 
-      const Comp = getComponentForPath(currentRoutePath)
-      return <Comp />
+      return (
+        <RoutePathContext.Provider value={routePath}>
+          {children ? (
+            children({
+              routePath,
+              getComponentForPath,
+            })
+          ) : (
+            <Comp />
+          )}
+        </RoutePathContext.Provider>
+      )
     }
   }
 )