+++ packages/react-static/src/static/webpack/rules/jsLoaderExternal.js
@@ -3,10 +3,7 @@ import babelPreset from 'babel-preset-react-static'
 export default function({ config }) {
   return {
     test: /\.(js|jsx|mjs)$/,
-    exclude: [
-      /@babel(?:\/|\\{1,2})runtime/,
-      ...config.babelExcludes || []
-    ],
+    exclude: [/@babel(?:\/|\\{1,2})runtime/, ...(config.babelExcludes || [])],
     use: [
       // 'thread-loader',
       {