+++ packages/react-static/src/utils/binHelper.js
@@ -1,3 +1,6 @@
+const path = require('path')
+const { escapeRegExp } = require('./')
+
 let ignorePath
 
 // Allow as much stack tracing as possible
@@ -8,7 +11,9 @@ require('@babel/register')({
     function babelIgnore(filename) {
       // true if should ignore
       return (
-        /\/node_modules\//.test(filename) ||
+        new RegExp(escapeRegExp(`${path.sep}node_modules${path.sep}`)).test(
+          filename
+        ) ||
         (ignorePath && ignorePath.test(filename))
       )
     },
@@ -63,6 +68,6 @@ process.on('unhandledRejection', r => {
 
 module.exports = {
   setIgnorePath(path) {
-    ignorePath = path ? new RegExp(path) : undefined
+    ignorePath = path ? new RegExp(escapeRegExp(path)) : undefined
   },
 }